tgbot/requirements.txt
aiogram==3.3.0
APScheduler==3.10.4
SQLAlchemy==2.0.29
python-dotenv==1.0.1
psycopg2-binary==2.9.9
uvicorn[standard]==0.29.0
fastapi==0.110.0
jinja2==3.1.3
pytz==2024.1
python-multipart==0.0.9
holidays==0.46
